Modified the seed file

+++ db/seeds.rb
@@ -1,5 +1,7 @@
 # Creation des "Plans" (type  d'abonnement)
-Plan.delete_all
+if Plan.count > 0
+  Plan.delete_all
+end
 tab = %w{ Standard Silver Gold }
 prices = [ 3990, 4990, 5990 ]
 tab.each_with_index do |e, index|
